# backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dashboard SOC API",
    "version": "0.1.0",
    "description": "SOC Ticketing System backend API. All responses are wrapped in a {success, message?, data?, error?, meta?} envelope.\n\nAuth: use Bearer JWT access tokens in the Authorization header.\n\nNotes:\n- /api/v1/auth/register is intended for SOC_MANAGER to create L1/L2 analysts.\n- Bootstrapping the first SOC_MANAGER account is out-of-band (DB seed/manual insert)."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "webhookApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "TicketSeverity": {
        "type": "string",
        "enum": ["low", "medium", "high", "critical"],
        "example": "high"
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "alice" },
          "password": { "type": "string", "example": "password123" }
        },
        "required": ["username", "password"]
      },
...
      "SuccessDashboardEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/Envelope" },
          {
            "type": "object",
            "properties": {
              "success": { "type": "boolean", "example": true },
              "data": { "$ref": "#/components/schemas/DashboardResponse" }
            }
          }
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "examples": {
              "badRequest": {
                "value": {
                  "success": false,
                  "error": { "code": "BAD_REQUEST", "message": "Invalid JSON body" }
                }
              }
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorEnvelope" },
            "examples": {
              "unauthorized": {
                "value": {
                  "success": false,
                  "error": { "code": "UNAUTHORIZED", "message": "Invalid token" }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Dashboard aggregates",
        "description": "Return role-based dashboard aggregates for the authenticated user. Use either range (e.g. 24h, 7d) or from/to (RFC3339) to define the activity window.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "example": "24h" },
            "description": "Activity window duration (supports Go duration like 24h or day shorthand like 7d). Ignored if from/to are provided. Default: 24h."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window start (RFC3339). Must be provided together with to."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window end (RFC3339). Must be provided together with from."
          },
          {
            "name": "recent_limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "example": 10 },
            "description": "Maximum number of recent tickets returned (1-50). Default: 10."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessDashboardEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "description": "Authenticate with username/password and receive access + refresh tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessTokenEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": ["Auth"],
        "summary": "Refresh access token",
        "description": "Exchange a valid refresh token for a new access token and new refresh token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessTokenEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "description": "Revoke a refresh token (server-side).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessageEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMeEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/change-password": {
      "post": {
        "tags": ["Auth"],
        "summary": "Change password",
        "description": "Change the current user's password. Revokes all refresh tokens after a successful change.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessageEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/users/{id}": {
      "patch": {
        "tags": ["Auth"],
        "summary": "Admin update analyst user",
        "description": "SOC_MANAGER-only. Update an analyst (L1/L2) user's full_name, username, role, and/or password.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminUpdateAnalystRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessageEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/BadRequest" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new analyst user",
        "description": "SOC_MANAGER-only endpoint to create L1/L2 analyst accounts.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessRegisterEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": { "$ref": "#/components/responses/BadRequest" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": ["Auth"],
        "summary": "List users",
        "description": "SOC_MANAGER-only. List users for administration.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessUsersListEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": ["Auth"],
        "summary": "Get user detail",
        "description": "SOC_MANAGER-only. Return user detail including sessions and ticket logs.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessUserDetailEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "Delete analyst account",
        "description": "SOC_MANAGER-only. Delete an existing L1/L2 analyst account.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessageEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/users/{id}/sessions/revoke": {
      "post": {
        "tags": ["Auth"],
        "summary": "Revoke user sessions (all or single)",
        "description": "SOC_MANAGER-only. Revoke refresh sessions of a target user. Omit session_id to revoke all active sessions, or provide session_id to revoke one active session.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RevokeUserSessionsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessRevokeUserSessionsEnvelope" },
                "examples": {
                  "revokeAllSuccess": {
                    "value": {
                      "success": true,
                      "message": "All active sessions revoked",
                      "data": {
                        "scope": "all",
                        "revoked_count": 2
                      }
                    }
                  },
                  "revokeAllNoChanges": {
                    "value": {
                      "success": true,
                      "message": "No active sessions were revoked",
                      "data": {
                        "scope": "all",
                        "revoked_count": 0
                      }
                    }
                  },
                  "revokeSingleSuccess": {
                    "value": {
                      "success": true,
                      "message": "Session revoked",
                      "data": {
                        "scope": "single",
                        "session_id": "019b2643-25c5-7c5b-8746-b388bedc739a",
                        "revoked_count": 1
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/tickets": {
      "get": {
        "tags": ["Tickets"],
        "summary": "List tickets",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10, "minimum": 1, "maximum": 100 }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketStatus" }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketSeverity" }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessTicketsListEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/tickets/{id}": {
      "get": {
        "tags": ["Tickets"],
        "summary": "Get ticket detail",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessTicketDetailEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/tickets/{id}/status": {
      "patch": {
        "tags": ["Tickets"],
        "summary": "Update ticket status",
        "description": "Update ticket status and persist a ticket log note.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTicketStatusRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessTicketStatusUpdatedEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/tickets/{id}/analyze": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Trigger external analysis",
        "description": "Triggers external AI analysis for a ticket using /api/v1/analyze/go contract, stores parsed analysis output to database, and records an audit log note.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AnalyzeTicketRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessAnalyzeTicketEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List notifications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 10, "minimum": 1, "maximum": 100 }
          },
          {
            "name": "is_read",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessNotificationsListEnvelope" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark notification as read",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessNotificationReadEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/notifications/stream": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Notifications SSE stream",
        "description": "Server-Sent Events stream. Emits 'notification' events with NotificationListItem JSON payload, and periodic 'ping' events.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "examples": {
                  "example": {
                    "value": "event: notification\ndata: {\"id\":\"...\",\"ticket_id\":\"...\",\"message\":\"New ticket...\",\"is_read\":false,\"created_at\":\"2026-03-04T18:24:22Z\"}\n\n"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/reports/tickets.csv": {
      "get": {
        "tags": ["Reports"],
        "summary": "Export tickets (CSV)",
        "description": "SOC_MANAGER-only export. Exports tickets in a time window including audit logs, IOCs, analysis, recommendations, and raw payload (nested fields are serialized as JSON columns).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "example": "7d" },
            "description": "Window duration (e.g. 24h, 7d). Ignored if from/to are provided. Default: 7d."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window start (RFC3339). Must be provided together with to."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window end (RFC3339). Must be provided together with from."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketStatus" }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketSeverity" }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 1000, "minimum": 1, "maximum": 10000 }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": {
              "text/csv": {
                "schema": { "type": "string" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/v1/reports/tickets.pdf": {
      "get": {
        "tags": ["Reports"],
        "summary": "Export tickets (PDF)",
        "description": "SOC_MANAGER-only export. Generates a PDF report for tickets in a time window.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "example": "7d" },
            "description": "Window duration (e.g. 24h, 7d). Ignored if from/to are provided. Default: 7d."
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window start (RFC3339). Must be provided together with to."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date-time" },
            "description": "Window end (RFC3339). Must be provided together with from."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketStatus" }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": { "$ref": "#/components/schemas/TicketSeverity" }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "format": "int32", "default": 1000, "minimum": 1, "maximum": 10000 }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF file",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  }
}